backend/requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
google-generativeai==0.3.2
python-dotenv==1.0.0
sqlalchemy==2.0.25
pydantic==2.5.3
python-multipart==0.0.6
alembic==1.13.1